Extract memory domain modules
