+++ src/memory-domain/index.js
@@ -1,0 +1,10 @@
+module.exports = {
+  observations: require('./observations'),
+  search: require('./search'),
+  context: require('./context'),
+  sessions: require('./sessions'),
+  recall: require('./recall'),
+  dedupe: require('./dedupe'),
+  compaction: require('./compaction'),
+  workspaces: require('./workspaces'),
+};
